requirements.txt
accelerate==1.6.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.6.0
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
einops==0.7.0
eval-type-backport==0.1.3
exceptiongroup==1.2.0
fairscale==0.4.13
filelock==3.13.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
hf_transfer==0.1.9
higher==0.2.1
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.6
importlib_metadata==8.6.1
importlib_resources==6.3.2
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.14.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
peft==0.9.0
pillow==10.2.0
portalocker==2.8.2
protobuf==5.26.0
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pynvml==12.0.0
pyparsing==3.1.2
pyproject==1.3.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.3
rich==13.7.1
safetensors==0.5.3
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentencepiece==0.2.0
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
sympy==1.12
threadpoolctl==3.4.0
timm==0.9.16
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
triton==2.2.0
trl==0.17.0
typing_extensions==4.10.0
tyro==0.8.3
tzdata==2024.1
urllib3==2.2.1
xformers==0.0.25.post1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
